+++ lib/std/textio-body.vhd
@@ -15,6 +15,9 @@
 --  along with this program; if not, see <http://www.gnu.org/licenses/>.
 -------------------------------------------------------------------------------
 
+library nvc;
+use nvc.polyfill.all;
+
 package body textio is
 
     procedure grow (l        : inout line;
@@ -723,17 +726,14 @@ package body textio is
 
     procedure write (l         : inout line;
                      value     : in real;
-                     justified : in side:= right;
+                     justified : in side := right;
                      field     : in width := 0;
-                     digits    : in natural:= 0 ) is
-        -- This may be compiled as VHDL-93 so cannot call TO_STRING directly
-        function fmt(value : real; spec : string) return string;
-        attribute foreign of fmt : function is "_std_to_string_real_format";
+                     digits    : in natural := 0 ) is
     begin
         if digits = 0 then
-            write(l, fmt(value, "%e"), justified, field);
+            write(l, to_string(value, "%e"), justified, field);
         else
-            write(l, fmt(value, "%." & integer'image(digits) & "f"),
+            write(l, to_string(value, "%." & integer'image(digits) & "f"),
                   justified, field);
         end if;
     end procedure;
@@ -751,8 +751,7 @@ package body textio is
                       justified : in    side  := right;
                       field     : in    width := 0) is
     begin
-        -- TODO
-        report "unimplemented" severity failure;
+        write(l, to_ostring(value), justified, field);
     end procedure;
 
     procedure hwrite (l         : inout line;
@@ -760,8 +759,7 @@ package body textio is
                       justified : in    side  := right;
                       field     : in    width := 0) is
     begin
-        -- TODO
-        report "unimplemented" severity failure;
+        write(l, to_hstring(value), justified, field);
     end procedure;
 
     function justify (value     : string;